+++ packages/kbn-utility-types/jest/package.json
@@ -1,3 +1,3 @@
 {
-  "types": "../target/jest/index.d.ts"
+  "types": "../target_types/jest/index.d.ts"
 }
