+++ packages/kbn-utility-types/src/jest/index.ts
