+++ packages/kbn-utility-types/src/tsd_tests/empty.d.ts
