+++ packages/kbn-utility-types/src/tsd_tests/test_d/values.ts
@@ -8,10 +8,21 @@
 
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { expectAssignable } from 'tsd';
-import { UnwrapPromise } from '../../index';
+import { Values } from '../..';
 
-type STRING = UnwrapPromise<Promise<string>>;
-type TUPLE = UnwrapPromise<Promise<[number, number]>>;
+// Arrays
+type STRING = Values<string[]>;
+type ASDF_FOO = Values<Array<'asdf' | 'foo'>>;
 
 expectAssignable<STRING>('adf');
-expectAssignable<TUPLE>([1, 2]);
+expectAssignable<ASDF_FOO>('asdf');
+expectAssignable<ASDF_FOO>('foo');
+
+// Objects
+type STRING2 = Values<Record<number, string>>;
+type FOO = Values<Record<number, 'foo'>>;
+type BAR = Values<{ foo: 'bar' }>;
+
+expectAssignable<STRING2>('adf');
+expectAssignable<FOO>('foo');
+expectAssignable<BAR>('bar');
