+++ packages/kbn-utility-types/src/tsd_tests/test_d/writable.ts
@@ -8,7 +8,7 @@
 
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { expectAssignable } from 'tsd';
-import { Writable } from '../../index';
+import { Writable } from '../..';
 
 type WritableArray = Writable<readonly string[]>;
 expectAssignable<WritableArray>(['1']);
