+++ src/core/server/saved_objects/service/lib/update_objects_spaces.test.ts
@@ -12,7 +12,7 @@ import {
   mockRawDocExistsInNamespace,
 } from './update_objects_spaces.test.mock';
 
-import type { DeeplyMockedKeys } from '@kbn/utility-types/target/jest';
+import type { DeeplyMockedKeys } from '@kbn/utility-types/jest';
 import type { ElasticsearchClient } from 'src/core/server/elasticsearch';
 import { elasticsearchClientMock } from 'src/core/server/elasticsearch/client/mocks';
